Use uintptr_t for addresses
Only 'uintptr_t' and 'intptr_t' are guaranteed to be of the same
size as a pointer. On Linux/Unix systems often 'unsigned long' and
'long' are used for this purpose. But on 64 bit Windows a 'long'
has only 4 bytes. So use 'uintptr_t' for addresses. Fall back to
'unsigned long' in case it is not available.
Also optimize the swath functions returning addresses by making
them inline.